+++ software_esp32/include/helper.h
@@ -50,4 +50,6 @@ String ip2String (IPAddress ipv4addr);
 String ConvBinUnits(size_t bytes, byte resolution);
 bool isNumber(const String& str);
 bool isFloat(const std::string& str);
+void replace (char* buffer,uint16_t size,char find, char with);
+